// apps/frontend/src/app/shared/models/enums/market-cap-thresholds.ts
export enum MarketCapThresholds {
  MegaCap = 200000000000, // More than $200 billion
  LargeCap = 10000000000, // $10 billion to $200 billion
  MidCap = 2000000000, // $2 billion to $10 billion
  SmallCap = 300000000, // ($300 million to $2 billion)
  MicroCap = 50000000, // ($50 million to $300 million)
}